+++ bindings/nodejs-old/types/event.ts
@@ -184,6 +184,13 @@ class TransactionProgressWalletEvent extends WalletEvent {
         super(WalletEventType.TransactionProgress);
         this.progress = progress;
     }
+
+    /**
+     * Returns the progress.
+     */
+    getProgress(): TransactionProgress {
+        return this.progress;
+    }
 }
 
 abstract class TransactionProgress {
@@ -292,4 +299,5 @@ export {
     SigningTransactionProgress,
     PerformingPowProgress,
     BroadcastingProgress,
+    TransactionProgressType,
 };