+++ bindings/python/examples/how_tos/accounts_and_addresses/create_account.py
@@ -18,12 +18,17 @@ if 'STRONGHOLD_PASSWORD' not in os.environ:
 secret_manager = StrongholdSecretManager(
     os.environ['STRONGHOLD_SNAPSHOT_PATH'], os.environ['STRONGHOLD_PASSWORD'])
 
-wallet = Wallet(os.environ['WALLET_DB_PATH'], client_options, coin_type, secret_manager)
+wallet = Wallet(
+    os.environ['WALLET_DB_PATH'],
+    client_options,
+    coin_type,
+    secret_manager)
 
 if 'MNEMONIC' not in os.environ:
     raise Exception(".env MNEMONIC is undefined, see .env.example")
 
-# Store the mnemonic in the Stronghold snapshot, this only needs to be done once
+# Store the mnemonic in the Stronghold snapshot, this only needs to be
+# done once.
 wallet.store_mnemonic(os.environ['MNEMONIC'])
 
 account = wallet.create_account('Alice')