+++ bindings/python/examples/how_tos/native_tokens/burn.py
@@ -19,13 +19,17 @@ if 'STRONGHOLD_PASSWORD' not in os.environ:
 wallet.set_stronghold_password(os.environ["STRONGHOLD_PASSWORD"])
 
 # Find native token with enough balance
-token = [native_balance for native_balance in balance.nativeTokens if int(native_balance.available, 0) >= 10][0]
+token = [
+    native_balance for native_balance in balance.nativeTokens if int(
+        native_balance.available,
+        0) >= 10][0]
 print(f'Balance before burning: {int(token.available, 0)}')
 
 burn_amount = 1
 
 # Send transaction.
-transaction = account.prepare_burn_native_token(token.tokenId, burn_amount).send()
+transaction = account.prepare_burn_native_token(
+    token.tokenId, burn_amount).send()
 print(f'Transaction sent: {transaction.transactionId}')
 
 # Wait for transaction to get included
@@ -33,5 +37,6 @@ blockId = account.retry_transaction_until_included(transaction.transactionId)
 print(f'Block included: {os.environ["EXPLORER_URL"]}/block/{blockId}')
 
 balance = account.sync()
-available_balance = int([native_balance for native_balance in balance.nativeTokens if native_balance.tokenId == token.tokenId][0].available, 0)
+available_balance = int(
+    [native_balance for native_balance in balance.nativeTokens if native_balance.tokenId == token.tokenId][0].available, 0)
 print(f'Balance after burning: {available_balance}')