+++ bindings/python/examples/wallet/12-prepare_output.py
@@ -5,7 +5,8 @@ import os
 
 load_dotenv()
 
-# In this example we will prepare an output with an address and expiration unlock condition and send it
+# In this example we will prepare an output with an address and expiration
+# unlock condition and send it.
 
 wallet = Wallet(os.environ['WALLET_DB_PATH'])