+++ bindings/python/iota_sdk/types/common.py
@@ -6,8 +6,6 @@ from typing import NewType
 
 HexStr = NewType("HexStr", str)
 
-HD_WALLET_TYPE = 44
-HARDEN_MASK = 1 << 31;
 
 class CoinType(IntEnum):
     IOTA = 4218
@@ -19,7 +17,8 @@ class CoinType(IntEnum):
 
 
 class Node():
-    def __init__(self, url=None, jwt=None, username=None, password=None, disabled=None):
+    def __init__(self, url=None, jwt=None, username=None,
+                 password=None, disabled=None):
         """Initialize a Node
 
         Parameters
@@ -42,7 +41,7 @@ class Node():
         self.disabled = disabled
 
     def as_dict(self):
-        config = {k: v for k, v in self.__dict__.items() if v != None}
+        config = {k: v for k, v in self.__dict__.items() if v is not None}
 
         if 'jwt' in config or 'username' in config or 'password' in config:
             config['auth'] = {}
@@ -73,7 +72,7 @@ class AddressAndAmount():
         self.amount = amount
 
     def as_dict(self):
-        config = {k: v for k, v in self.__dict__.items() if v != None}
+        config = {k: v for k, v in self.__dict__.items() if v is not None}
 
         if 'amount' in config:
             config['amount'] = str(config['amount'])