+++ bindings/python/iota_sdk/utils.py
@@ -16,6 +16,7 @@ from dacite import from_dict
 if TYPE_CHECKING:
     from iota_sdk.types.block import Block
 
+
 class Utils():
 
     @staticmethod
@@ -101,7 +102,8 @@ class Utils():
         })
 
     @staticmethod
-    def compute_foundry_id(alias_id: HexStr, serial_number: int, token_scheme_kind: int) -> HexStr:
+    def compute_foundry_id(alias_id: HexStr, serial_number: int,
+                           token_scheme_kind: int) -> HexStr:
         """Computes the foundry id.
         """
         return _call_method('computeFoundryId', {
@@ -145,7 +147,8 @@ class Utils():
         }))
 
     @staticmethod
-    def compute_token_id(alias_id: HexStr, serial_number: int, token_scheme_kind: int) -> HexStr:
+    def compute_token_id(alias_id: HexStr, serial_number: int,
+                         token_scheme_kind: int) -> HexStr:
         """Computes a token id from the alias id, serial number and token scheme type.
         """
         return _call_method('computeTokenId', {
@@ -171,7 +174,8 @@ class Utils():
         })
 
     @staticmethod
-    def verify_ed25519_signature(signature: Ed25519Signature, message: HexStr) -> bool:
+    def verify_ed25519_signature(
+            signature: Ed25519Signature, message: HexStr) -> bool:
         """Verifies an ed25519 signature against a message.
         """
         return _call_method('verifyEd25519Signature', {
@@ -180,7 +184,8 @@ class Utils():
         })
 
     @staticmethod
-    def verify_secp256k1_ecdsa_signature(public_key: HexStr, signature: HexStr, message: HexStr) -> bool:
+    def verify_secp256k1_ecdsa_signature(
+            public_key: HexStr, signature: HexStr, message: HexStr) -> bool:
         """Verifies a Secp256k1Ecdsa signature against a message.
         """
         return _call_method('verifySecp256k1EcdsaSignature', {