+++ bindings/python/CHANGELOG.md
@@ -57,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `Wallet::{generate_mnemonic(), verify_mnemonic()}` since they're available from `Utils`;
+- `HD_WALLET_TYPE`, `HARDEN_MASK` constants;
 
 ### Fixed